Refactor code related to collider and non-collider triples
This commit moves the addition of new triples into their respective conditional blocks to ensure proper collection. This refactoring helps in organizing the logic by adding the triple in line with the condition that resulted in its creation, rather than adding it indiscriminately afterward. This also keeps the list of triples specific to their conditions, leading to cleaner and more efficient code.