Simplify ForkJoinPool utilization
The commit replaces the use of custom-generated ForkJoinPool with ForkJoinPool.commonPool() in VarianceVectorForkJoin and RealCovarianceMatrixForkJoin classes. This simplifies the code structure and enhances readability. It also puts an upper limit on the number of threads in RealCovarianceMatrixForkJoin to the number of columns if initially assigned more. Resultantly, it abandons the lengthy exception handling and waiting sessions for pool quiescence, calling shutdown() instead.