Add new dependency-reduced-pom.xml in tetrad-gui module
Implemented a new Maven POM file for the tetrad-gui module to specify its dependencies and build configuration. It sets the Java version, specifies the wagon-ssh extension and maven-compiler-plugin, and configures how the resource files will be processed. The Maven Shade plugin is also configured to create a shaded JAR file.