Refactor code and update variable names in IcaLingD algorithm.
The code refactor involved rewriting comments and renaming some variables to make them more understandable in the IcaLingD and associated classes. The variable "thresholdSpine" has been renamed to "thresholdW". Further, a new variable "wThreshold" has been introduced in IcaLingD for setting the threshold value for the W matrix. Accompanied with this, the method 'setWThreshold' has been implemented to handle the setting of this new threshold. Lastly, the verbose flag for IcaLingD's 'search' method is revised to only print all stable and unstable graphs if enabled.