Refactor check for latent variable algorithm
The method for checking if an algorithm is a latent variable algorithm is refactored out of GraphCard.java into GraphSearchUtils.java as a static method. This improves code maintainability as the method can now be readily used wherever needed. It also reduces the complexity of the GraphCard class by removing unnecessary responsibilities.