Refactor IdaModel and IdaEditor for more efficient data handling and presentation
The IdaModel and IdaEditor classes have been refactored to better handle and present estimated and actual data. In IdaModel, adjustments were made to accurately track and retrieve the estimated and true CPDAGs using IdaCheck objects, while in IdaEditor, the table presentation of these data types was improved. Excessive and redundant code has also been trimmed as part of the changes.