Refactor IdaCheck to exclude private DataSet attribute
The DataSet attribute has been removed from the IdaCheck class in the Tetrad library. As a result, the Ida instance only takes three parameters: the dataSet, graph, and nodes. The dataSet is directly passed when instantiating Ida which streamlines the codebase. Moreover, specific import statements replace a generic one for better clarity and efficiency.