Update Ida algorithm to calculate total effects
The Ida search algorithm has been updated to calculate the total and absolute total effects of a variable on another variable. The change is reflected in the method which reports the sorted regression of possible parents of a variable on other variables. Regressor calculations were edited for better clarity and efficiency. Other classes and files have been modified accordingly.