Add setParallelism method to Comparison class
A method to set the parallelism level, setParallelism, has been added to the Comparison class in `Algorithm Compare`. This method allows the user to specify the number of threads for the function to run. Additionally, the GUI model `AlgcomparisonModel.java` was updated to set the degree of parallelism to 1 by default.