Add COMPARE_GRAPH_ALGCOMP parameter to various simulations
The parameter COMPARE_GRAPH_ALGCOMP has been added to various simulation files in the Tetrad library. This parameter allows us to compare different types of graphs in algorithm comparison. It has been also added to the index.html file for proper documentation. Furthermore, code to handle this parameter has been incorporated in the AlgcomparisonModel class in the Tetrad GUI module.