Remove COMPARE_GRAPH_ALGCOMP parameter addition from multiple simulations
This commit removes the newly added parameter 'COMPARE_GRAPH_ALGCOMP' from simulation files in the Tetrad library. The parameter was initially included for comparing graphs during algorithm comparison but is deemed unnecessary. The associated GUI handling in the AlgcomparisonModel class, and its documentation in the index.html file, has been removed as well.