Update graph loading and saving functionality with amat.pag format
The existing graph loading functionality was updated to support the "amat.pag" format of PCALG. The class names and some of the method names have been renamed to reflect this change. Additionally, an option to save graphs in the "amat.pag" format was also introduced. The original PCALG option in saving graphs was commented out since the "amat.pag" and the "amat.cpdag" formats seem to be more explicit and accurate.