Implement Meek Rules in Paths class
A new import, `edu.cmu.tetrad.search.utils.MeekRules`, was added to the Paths class. This was used to implement Meek rules which aid in orienting edges in the graphical model. The equals_check was adjusted - now it also verifies the maximality by comparing the original graph with the newly oriented one. If both are equal, it returns true; otherwise, it returns false.