Refactor graph methods and improve documentation
The import statement for GraphInPag in Paths.java has been removed as it's unnecessary. The comments/documentation for the method "anteriority" in Paths.java and GraphUtils.java has been improved for clarity, as well as additional documentation added for the methods "getGraphWithoutXToY" and "anteriority" in GraphUtils.java. This should provide a clearer understanding of these methods.