Add new functions for graph manipulation
This update adds the functionality to run final FCI (Fast Causal Inference) rules and checks for Maximal Partially Directed Acyclic Graph (MPDAG) rules on a graph. Also adds undo and reset to original options in the graph workbench for better manipulation of graphs. The code refactors and centralises the conjoint menu insertion and highlight items set-up for a more streamlined code structure.