Refine graph handling methods
We've improved the handling of graph methods in the code. Changes include initializing the LinkedList `graphStack` as final, removing data from `modelNodesToDisplay` and `modelEdgesToDisplay` when corresponding nodes and edges are removed, and adding extra checks for `SemGraph` in `AbstractWorkbench`. In `GraphWrapper`, the `setGraph` method now directly adds the graph without converting it to an `EdgeListGraph`, and a similar change has been made in `SemGraphWrapper`.