Refactor PAG coloring and adjust paste naming in GUI
The PAG coloring process was refactored in different classes (GraphEditor, GraphCard, SemGraphEditor) to centralize it within the GraphUtils class. In addition, the naming for the paste action within the PasteSubgraphAction was adjusted from "Paste Selected Graph" to "Paste Selected Items" for better accuracy. In addition, a new class for PAG Edge Type Instructions was introduced.