Update graph check dialogs to use containing JScrollPane
The dialogs that pop up to check if a graph is a particular type (like MPAG, CPDAG, etc.) have been updated. They now get displayed in the JScrollPane that contains the workbench currently in use. This makes it more intuitive for the user to correlate the dialog message with the related graph on the workbench.