Update graph edge explanation and help window size
The explanation for graph edges types in `graph_edge_types.html` has been expanded, providing more specific details when PAG coloring is turned on. Changes include clarifications of what solid and dashed edges represent as well as the meaning of thickened edges. In addition, the height of the help window in `TetradHelp.hs` has been increased for better user experience.