Add functionality for latent confounder detection in bidirectional edges
The commit includes the addition of a new method in GraphUtils.java to assess if a bidirected edge has a latent confounder in the true graph. This method throws an IllegalArgumentException if the edge is not bidirected. Furthermore, statistics classes BidirectedLatentPrecision and NumCorrectBidirected have been updated to utilize this method. Search configurations in LvLite.java have also been modified to improve error detection and handling.