Add verbosity checks to logger in FciOrient and LvLite
This commit adds checks for verbosity before logging messages in the classes FciOrient and LvLite. Now, messages regarding graph edge orientations within the FciOrient class and operations related to unshielded colliders and edge orientations in the LvLite class will only be logged if verbosity is enabled. These changes help to prevent unnecessary logging when verbose mode is not active.