Update LvLite class documentation
The documentation for the LvLite class has been updated to correctly reflect its purpose. It accurately explains that LvLite is an implementation of the LV algorithm for learning causal structures from observational data, using a combination of independence tests and scores to search for the best graph structure given a data set and parameters.