Improve reachability algorithm performance in CPDAGs and PAGs
The update introduces an enhancement in the reachability algorithm to work more efficiently with CPDAGs and PAGs. The aim is to generate virtual edges directed towards the arrow, enabling the reachability algorithm to find any implied colliders along the path. In addition, minor alterations have been made to the 'TestGraphUtils.java' file to increase the number of random graph generations.