Rename LvLite to BfciSb and refactor algorithm implementation
The LvLite class has been renamed to BfciSb in both the main algorithm and the corresponding algorithm comparison files. The algorithm implementation has been significantly refactored to match the new BFCI-SB (BFCI Score-based) algorithm. This includes updates to the method of scoring and edge orientation, as well as changes to how certain steps are handled in the algorithm.