Update graph functionality and improve path string representation
This commit enhances the functionality of the graph and path representation, especially for the adjustment sets. It changes the GraphUtils path string method to include a boolean showBlocked parameter to show if a path is blocked. It also introduces a new JTextFieldWithPrompt class which extends JTextField to include a prompt in the text field. This is used to allow users to enter conditioning variables for paths in graph. Lastly, several changes were made to clean up and enhance the UI in PathsAction.