Update Paths class to prevent adding duplicate paths
Modified the implementation in Paths.java to prevent adding duplicate paths to the 'paths' list. Prior to adding a new path to the list, the code now checks if the path already exists in the list. Also, commented out some unnecessary conditional checks that were triggering the addition of more paths.