Improved updates for independence and dependence counts in MarkovCheckEditor
Reworked the rowIndex check in getValueAt method to prevent out of range errors. Added a timer to call updates every second to "# independencies" and "# dependencies" labels in MarkovCheckEditor. Created a new interface, ModelObserver, and integrated observer pattern in MarkovCheck to notify observers of changes. These measures ensure that data is updated in real-time and errors related to out of range indices are avoided.