Refactor BfciSb implementation and add a GRaSP algorithm option
The BfciSb class was significantly refactored. A GRaSP algorithm option was added which can be used to get a CPDAG like GFCI with FGES. Moreover, the reorientWithCircles method, which reorients all edges in a graph following the PAG structure, was moved higher up in the class structure. The order of calling methods in the main search procedure has also been tweaked.