Add 'allowTucks' option to LV-Lite search algorithm
Added a new private variable 'allowTucks' in the LV-Lite class. This variable serves as a flag to enable or disable 'tucks' in the search procedure. Also, functions for setting this variable and modifying its behavior throughout the algorithm were implemented. The relevant user parameter has also been added in the algorithm comparison and documentation files.