Refactor collider orientation and remove unused parameters
The collider orientation logic in the LvLite.java and LvLiteDsepFriendly.java files is refactored and made more concise. Unnecessary parameters and methods related to depth settings for search algorithms in these files are also removed. The default value for 'doDiscriminatingPathColliderRule' is now set to true in the manual index.html file.