Add node selection feature to MarkovCheckEditor and remove parameters from LvLiteDsepFriendly
In the MarkovCheckEditor class, a new selection box is added to allow the user to select specific nodes from the table view. Additionally, the code has been modified to sort the names of nodes in a specific order. In the LvLiteDsepFriendly.java, some parameters related to GRaSP have been removed to simplify the structure. These parameters seemed to no longer be in use.