Refactor LvLite.java: Add option to choose initial CPDAG algorithm, remove unused methods
The LvLite.java class has been updated to include an option to choose the initial CPDAG algorithm. The options are now BOSS and GRASP. Methods that were previously unused or duplicated in this file have been removed for improved code clarity and efficiency. The reorientation of edges and determination of the final orientation of the graph have been restructured and moved to improve readability. Other minor code restructuring and reordering have also been done for better organization of the class functions.