Add resultsPath field to GridSearchModel
A new field, resultsPath, has been added to the GridSearchModel class, along with its setter and getter methods. This field represents the path to the result folder. This is set after a comparison has been run, and used to add additional contents like simulation, algorithm, table columns, and verbose output to the comparison results.