Update Maven plugin versions and add documentation comments
The commit includes updating versions of the `maven-compiler-plugin` and `maven-shade-plugin` and changes the `java.version` property in several pom.xml files. Alongside these updates, it also includes adding a significant amount of JavaDoc comments in multiple class files to improve code readability and understanding. The commit also removes an unused variable in the `FaskForbiddenGraphModel.java`.