Reduce histogram bins and optimize variable removal in Markov check
The number of bins in the histogram has been reduced from 20 to 10 for compression. Moreover, an optimization has been added to the Markov check algorithm to remove extraneous variables more efficiently. The check for variable removal will now be performed in the specific sections of the algorithm where it is necessary.