Implement prevention of graph cycles in Meek rules
The Meek rules now include a provision to prevent cycles in the graph. The option to prevent cycles can be toggled on by the user. A check has been added to ensure that new cycles are not created while adding arbitrary unshielded colliders. The colliderAllowed method has been updated to include a check on graph ancestors. This prevents colliders from being activated where they would introduce cycles, thus maintaining the acyclic nature of the graph.