Refactor logging and collider checks in MeekRules and PcCommon
Revised logging to be controlled by a flag and added explanatory comments for increased readability. In MeekRules, logger is used to prevent cycles while orienting edges. In PcCommon, the method 'colliderAllowed' was moved to a more appropriate location and all the logging in the file was modified to use the updated log method.