Update repairFaultyPag method and enhance logging in GraphUtils
The 'repairFaultyPag' method in GraphUtils has been updated to not return a graph. Now, it directly modifies the graph input parameter. Additionally, this change was propagated to the classes that call this method: Bfci, Fci, GraspFci, Gfci, LvLite, and SpFci. Also, logging details have been added in GFci.java to indicate the start and end of various stages in the process.