Update graph algorithm and remove unnecessary debug log
In this commit, several changes were made to the graph algorithm in various files, especially in classes SemGraph, LvLite and RegressionDataset. Pointless system.out calls used for debug log were removed. Some variables in LvLite were changed to improve the accuracy of the algorithm. Also, several redundant blocks of code were eliminated, increasing the simplicity and readability of the code. Meanwhile, a double-check method for DDP construct was added to ensure the fidelity of the construct.