Rename and modify NumInducedAdjacenciesInPag to NumEdgeInEstInTrue
The implementation of the NumInducedAdjacenciesInPag class has been renamed to NumEdgeInEstInTrue to clarify the function of the class. This change is in line with the alteration in the logic of determining induced adjacency: instead of looking for adjacencies in the estimated graph but not in the true graph, the code now checks for adjacencies which are present in both the estimated and true graphs. Also, minor changes have been made in GridSearchEditor and NumCoveringAdjacenciesInPag class, adjusting the sequence of method calls and the abbreviation used respectively.