Refactor LvLite and update search methods
The LvLite class has been refactored to encompass modifications in search methods and other configurations for creating a PAG. New methods have been created to handle instances like 'BossSearch', 'GraspSearch', and 'FciOrient' in a more organized and cleaner way. Furthermore, an improvement with 'allPaths' method in the Paths class has been extended to limit the maximum number of paths. Lastly, the BFci and GraspFci classes were updated to use 'SepsetsMinP' instead of 'SepsetsGreedy'.