Refactor LvLite and update FciOrient; add tucking and testing flags
The LvLite class has been refactored to improve the efficiency of the overall algorithm. A number of additions and modifications were made to improve existing functions and add new ones. The FciOrient class was also updated to correct potential issues with path length. New options were introduced to allow control over the tucking and testing stages in the search algorithm.