This version of LV-Lite is regularly passing a Markov check.
Refactor LvLite algorithm and associated files

The LvLite algorithm has been refactored to simplify the code and improve readability. Unnecessary features, such as the EXTRA_EDGE_REMOVAL_STEP, which added complexity to the code, have been removed. The code related to edge processing and graph generation has also been significantly modified to be more efficient and provide cleaner orientation of the graph. This revision ensures the orientation process is more streamlined and less prone to errors.