Add new SepsetFinder class and refactor related classes
A new class, SepsetFinder, has been added. This class is responsible for finding a Sepset between two given nodes, which is a complex task. Multiple methods that were originally located in the Paths class have been moved to the new SepsetFinder class. This was done to improve modularity and maintainability by separating concerns. Code in the FciOrient class has been updated to use the new class.