Add depth parameter to sepset finding functions
The 'depth' parameter has been added as an argument to 'getSepSet' functions in multiple files to provide more control over search depth. The detailed code changes also extend this parameter to 'isUnshieldedCollider' and 'getSepSetContaining' functions. These changes are based on the observation that including 'depth' in the sepset functions provides the possibility of specifying the maximum number of conditioning variables for independence tests, thus improving the flexibility and efficiency of the algorithm.